fix(review): add changeset and extract deduplicatePlayers to shared utility
