+++ .changeset/fix-libero-deduplication.md
@@ -1,0 +1,5 @@
+---
+'volleykit-web': patch
+---
+
+Fixed duplicate libero players in scoresheet parsing and crop button visibility on iOS
